Moved New Features and Migration Guide to seperate docs
Getting started contained a section for Getting Started and a Section for Migration guide.

This pull request moved those sections (almost) verbatim to new corresponding files.

The changes to the actual content only contains title updates (to give context now that the section is in its own file) and for the New Feature content was updated by replacing 3.x with 3.0